+++ engine/include/image_component.hpp
@@ -9,15 +9,15 @@ namespace engine {
 		private:
 			std::pair<double, double> m_position_relative_to_object = std::make_pair(0, 0);
 		protected:
-			int zoom_factor;
+			double zoom_factor;
 			SDL_Rect canvasQuad;
 		public:
 			ImageComponent();
-			ImageComponent(GameObject &game_object, std::string image_path, int zoom_factor);
+			ImageComponent(GameObject &game_object, std::string image_path, double zoom_factor);
 			ImageComponent(
 				GameObject &game_object,
 				std::string image_path,
-				int zoom_factor,
+				double zoom_factor,
 				std::pair<double, double> position_relative_to_object);
 			void Init();
 			void Draw();
